Added proper command-line option parsing.
User can now specify which individual checks they wish to do, if they only want to do a few of the footprints, or, for instance, skip the Windows registry since it takes so long. If no options are specified, it just runs all footprints, as one would expect.

Also, the package name is no longer included in the script invocation. It was not used anywhere in the script, so it was never necessary to specify.